question_answering_from_pdf/requirements.txt
langchain
pypdf
unstructured
sentence_transformers
pinecone-client
llama-cpp-python
huggingface_hub
streamlit
sqlite3
nnltk
python-dotenv
